Fix error in action controller has rails handler (#141)
* Pass rescued errors into has_rails_handler?

It looks like has_rails_handler? was supposed to take in an `exception`
param. It referred to an exception method/local variable, but we were
running into NoMethodErrors. This passes the rescued exception into the
method so that we can use it.

* Use the block form of Logger#debug

This avoids building the log string if the log level is not debug.