feat: Add support for latest Dalli versions
The `Dalli::Server` class has become deprecated.

From `3.0.0` up until `3.1.12` the same logic is in
`Dalli::Protocol::Binary`.

Starting with `3.1.12`, the request handling from
`Dalli::Protocol::Binary` has been moved to `Dalli::Protocol::Base`.

The `@hostname` and `@port` fields are dropped between in `v3.1.1`,
and only the `:hostname` and `:port` accessors  are available,
but those have been available even earlier, definitely
in the `2.0.0`, which is the oldest supported.

Signed-off-by: Ferenc Géczi <ferenc.geczi@ibm.com>