fix: Wrong number of arguments redis_client.rb:424:in `pipelined'
The breaking change that we have to adapt to is a change from
`redis-client-0.22.0` to `redis-client-0.22.1`,
namely the signature of `pipelined` changed from
`def pipelined` to `pipelined(exception: true)`.

Now, that only Ruby 3 is supported, we can focus on explicitly
delegating keyword arguments, which cover all future
keyword and non-keyword argument changes.

Signed-off-by: Ferenc Géczi <ferenc.geczi@ibm.com>